recommend/requirements.txt
Flask==3.0.2
pymysql==1.1.0
Werkzeug==3.0.1
python-dotenv==1.0.1
apscheduler==3.10.4
numpy==1.26.3
pandas==2.0.3
scikit-learn==1.2.2